feat(macos): consolidate playlist surfaces — port multi-select, delete+undo, drop-to-add into PlaylistView (#985)
The rich PlaylistDetailView (#540) was built but never routed; the routed
PlaylistView had the collage hero + inline rename but none of the track
interactions. Rather than swap (regressing editing) or delete (losing the
interactions), port the interactions into the routed screen and retire
the dead one:

- TrackRow grows the same optional isSelected/onSelect multi-select API
  TrackListRow got in #217 — defaulted, so AlbumDetailView / SearchView
  call sites are untouched. Selection rail, tint, and a11y traits match.
- The #217 gesture plumbing (SelectionClickModifier, FavoriteTapShield)
  moves from TrackListRow privates to Components/TrackSelectionGestures
  so both row families share one arbitration; a PlayTapWhenUnselectable
  modifier preserves TrackRow's bare-tap play on the single-select path.
- PlaylistView gains: Cmd/Shift multi-select resolved in display space
  (so Shift ranges match the ⌘F-filtered list, while playback indexes the
  true playlist position), Delete/Backspace batch remove with the
  10-second undo toast, Esc-clears-selection, a transport-bar
  'N selected / Remove' affordance, and drop-to-add with the accent
  drop-target border. Selection clears on playlist switch and filter
  edits so hidden rows are never silently batch-removed.
- removeFromPlaylist / undoRemoveFromPlaylist re-key off
  playlistTracks[playlistId] (the cache PlaylistView renders from)
  instead of the currentPlaylistTracks mirror, which stays in step when
  it was showing the mutated playlist. addToPlaylist's post-add re-fetch
  is now unconditional — its only caller is the on-screen drop handler.
- PlaylistDropPayload and UndoRemovalToast move to Components/;
  PlaylistDetailView.swift (732 lines, zero call sites) is deleted and
  the stale doc references updated. PlaylistDetailInteractionTests
  renamed to PlaylistInteractionTests, contracts unchanged.

Closes #985. Needs an interactive QA pass: selection gestures, Delete +
undo round-trip, drop-to-add, and the focusable(true) Tab interplay on
the playlist screen.

Co-developed against the #217 Library implementation as the reference
pattern.
